requirements.txt
openai-whisper
tqdm
colorama
pydub
ffmpeg-python
